Abstract most of Detail Pager Fragment with all the playback controls so that other screens that want to show paging content can easily extend
